a simple cli
